[ty] Split reachability analysis into two files
